Update trampoline feature bit
The `option_scid_alias` feature (https://github.com/lightning/bolts/pull/910)
is going to clash with the feature bit we use for trampoline payments.

We are only using this bit to signal trampoline support in invoices, so
we now set two bits for trampoline in invoices (the new one and the old one
for backwards-compatibility).

We support reading old invoices that only set the previous bit to allow
for a transition period.